+++ app/services/carto/user_metadata_export_service.rb
@@ -93,6 +93,8 @@ module Carto
       api_keys = exported_user[:api_keys] || []
       user.api_keys += api_keys.map { |api_key| Carto::ApiKey.new_from_hash(api_key) }
 
+      user.static_notifications = Carto::UserNotification.create(notifications: exported_user[:notifications])
+
       # Must be the last one to avoid attribute assignments to try to run SQL
       user.id = exported_user[:id]
       user
@@ -170,8 +172,7 @@ module Carto
 
       user_hash[:rate_limit] = export_rate_limit(user.rate_limit)
 
-      # TODO
-      # Organization notifications
+      user_hash[:notifications] = user.static_notifications.notifications
 
       user_hash
     end