+++ spec/services/carto/user_metadata_export_service_spec.rb
@@ -23,6 +23,7 @@ describe Carto::UserMetadataExportService do
   def create_user_with_basemaps_assets_visualizations
     create_account_type_fg('FREE')
     @user = FactoryGirl.create(:carto_user)
+    @user.static_notifications = Carto::UserNotification.new(notifications: full_export[:user][:notifications])
     @map, @table, @table_visualization, @visualization = create_full_visualization(@user)
 
     @tiled_layer = FactoryGirl.create(:carto_tiled_layer)
@@ -84,6 +85,12 @@ describe Carto::UserMetadataExportService do
   describe 'import latest' do
     it 'imports correctly' do
       import_user_from_export(full_export)
+    end
+  end
+
+  describe 'import v 1.0.3' do
+    it 'immports correctly' do
+      import_user_from_export(full_export_one_zero_three)
     end
   end
 
@@ -406,6 +413,7 @@ describe Carto::UserMetadataExportService do
       expect(st1.retrieved_items).to eq st2['retrieved_items']
       expect(st1.state).to eq st2['state']
     end
+    @imported_user.static_notifications.notifications.should eq full_export[:user][:notifications]
   end
 
   def full_export_import_viewer(path)
@@ -669,7 +677,14 @@ describe Carto::UserMetadataExportService do
             created_at: DateTime.now,
             updated_at: DateTime.now
           }
-        ]
+        ],
+        notifications: {
+          builder: {
+            onboarding: true,
+            :"layer-style-onboarding" => true,
+            :"layer-analyses-onboarding" => true
+          }
+        }
       }
     }
   end
@@ -681,6 +696,10 @@ describe Carto::UserMetadataExportService do
   end
 
   let(:full_export_one_zero_two) do
-    full_export.except(:rate_limit)
+    full_export_one_zero_three.except(:rate_limit)
+  end
+
+  let(:full_export_one_zero_three) do
+    full_export.except(:notifications)
   end
 end