+++ NEWS.md
@@ -922,6 +922,7 @@ More information at [Dropbox migration guide](https://www.dropbox.com/developers
 * Avoid exporting orphan raster overviews in user migrator
 * Set `soft_geocoding_limit` to default to false.
 * Do not export local visualizations lacking a map
+* Add notifications to user migrator (#13844)
 * Docs, fixed incorrect grammar in en.json file (customer reported).
 
 ### NOTICE